--- a/go.tools.mod
+++ b/go.tools.mod
@@ -5,6 +5,9 @@ go 1.14
 require (
 	github.com/jstemmer/go-junit-report v0.9.1 // indirect
 	github.com/mattermost/mattermost-utilities/mmgotool v0.0.0-20200721093743-053c38dcd293 // indirect
+	github.com/philhofer/fwd v1.0.0 // indirect
 	github.com/reflog/struct2interface v0.6.1 // indirect
+	github.com/tinylib/msgp v1.1.2 // indirect
+	github.com/ttacon/chalk v0.0.0-20160626202418-22c06c80ed31 // indirect
 	github.com/vektra/mockery v1.1.2 // indirect
 )