We use the tinylib/msgp serializer which generates methods for fast
serialization and de-serialization. Because of this, we can get rid of the
reflect overhead which comes with generic serializer.

However, the library is a cumbersome to work with. Because it does not support package
level parsing, we need to move some code temporarily for the generation to work
and then move them back.

We choose model.User and model.Session as our candidates for this migration.

model.User is by far our hottest cache struct because of getUserStatusesByIds calls.
Of all the cache calls, model.User is nearly 90% of them. Following the 80-20 rule,
this should be the only struct good enough to migrate. However, we add one more.

model.Session is actually the 4th hottest struct. But it is called very frequently
during user logins, as found from CPU profiles during load testing.

```
name            old time/op    new time/op    delta
LRU/User=new-8    19.2µs ± 1%    13.6µs ± 1%  -28.75%  (p=0.000 n=10+10)

name            old alloc/op   new alloc/op   delta
LRU/User=new-8    9.93kB ± 0%    9.79kB ± 0%   -1.32%  (p=0.000 n=9+10)

name            old allocs/op  new allocs/op  delta
LRU/User=new-8       102 ± 0%       101 ± 0%   -0.98%  (p=0.000 n=10+10)

name               old time/op    new time/op    delta
LRU/Session=new-8    7.35µs ± 7%    3.47µs ± 6%  -52.75%  (p=0.000 n=10+10)

name               old alloc/op   new alloc/op   delta
LRU/Session=new-8    2.92kB ± 0%    1.71kB ± 0%  -41.59%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
LRU/Session=new-8      36.0 ± 0%      28.0 ± 0%  -22.22%  (p=0.000 n=10+10)
```

Comparison of pooled with un-pooled with only sessions.

```
name \ time/op     master.txt   new-unpooled.txt  new-pooled.txt
LRU/Session=new-8  7.35µs ± 7%       3.59µs ± 9%     3.47µs ± 6%

name \ alloc/op    master.txt   new-unpooled.txt  new-pooled.txt
LRU/Session=new-8  2.92kB ± 0%       2.15kB ± 0%     1.71kB ± 0%

name \ allocs/op   master.txt   new-unpooled.txt  new-pooled.txt
LRU/Session=new-8    36.0 ± 0%         29.0 ± 0%       28.0 ± 0%
```

This is still a micro-optimization at a high-level as is apparent from the
load test results attached. That is because the other vmihailenco/msgpack API
used by the cool structs doesn't allow for a alloc-free implementation like msgp
does. So a good amount of CPU time is just spent in mark and sweep. MM-27457 tracks
that effort.